Extract GPS decoding logic externally
