+++ ublox/Cargo.toml
@@ -16,6 +16,12 @@ ubx_proto27 = []
 ubx_proto31 = []
 default = ["std", "serde", "ubx_proto23"]
 
+# Unlock RXM-SFRBX GPS and QZSS interpretation
+sfrbx-gps = [
+    "gnss-protos",
+    "gnss-protos/gps",
+]
+
 alloc = []
 std = []
 
@@ -31,6 +37,8 @@ serde = { version = "1.0", optional = true, default-features = false, features =
     "derive",
 ] }
 ublox_derive = { path = "../ublox_derive", version = "=0.2.0" }
+
+gnss-protos = { version = "0.0.1", optional = true }
 
 [dev-dependencies]
 cpu-time = "1.0"
