--- ublox/src/ubx_packets/packets/rxm_sfrbx/gps/unscaled/frame5/mod.rs
@@ -1,5 +1,0 @@
-mod page_1_24;
-pub(crate) use page_1_24::*;
-
-mod page_25;
-pub(crate) use page_25::*;
