+++ ublox/tests/rxm_sfrbx/mod.rs
@@ -1,1 +1,2 @@
+#[cfg(feature = "sfrbx-gps")]
 mod gps;
